fix a few bugs. refactor tests
